# docs/api/openapi-full.yaml
openapi: 3.1.0
info:
  title: DataSafeS3 REST Admin API
  version: 1.1.0
  summary: Complete JSON REST API served at /api/v1 (excluding S3 XML)
  description: |
    **Full hand-written OpenAPI 3.1** for all Admin JSON routes registered in internal/api/server.go.

    ## Authentication
    | Scheme | Use |
    |--------|-----|
    | **BearerAPIToken** | Integration tokens ds_* (recommended for automation) |
    | **BearerJWT** | JWT from POST /admin/login or OIDC callback (console sessions) |

    Public endpoints: GET /health, POST /admin/login, OIDC bootstrap, GET /public/share/*.

    ## Not included
    - GET /metrics (Prometheus text)
    - S3 XML API on paths outside /api/v1 (SigV4, port 9000)

    ## Errors
    Standard envelope: JSON object with error field (HTTP 4xx/5xx).

    Interactive docs for **Community subset**: GET /api/v1/docs (Swagger UI).

    Canonical file: [openapi-full.yaml](../../api/openapi-full.yaml)
  contact:
    name: DataSafeS3 / Ilya Trachuk
    url: https://github.com/DirektorBani/DataSafeS3
  license:
    name: Apache 2.0
    identifier: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Via Caddy (console origin)
  - url: http://localhost:9000/api/v1
    description: storage-server direct
security:
  - BearerAPIToken: []
  - BearerJWT: []
components:
  securitySchemes:
    BearerAPIToken:
      type: http
      scheme: bearer
      bearerFormat: ds_
      description: |
        Integration API token with prefix ds_. Create in console: Access → API tokens.
        Inherits creator's role and bucket access.

    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT from POST /admin/login, MFA completion, or OIDC callback.
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        mfa_required: { type: boolean }
        mfa_token: { type: string }
        mfa_setup_required: { type: boolean }
        scheduled_delete_at: { type: string, format: date-time }
        trashed: { type: boolean }
        trash_id: { type: string }
    FederationCluster:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        endpoint: { type: string }
        region: { type: string }
        status: { type: string }
        cluster_id:
          type: string
          description: Trusted cluster id this federation peer belongs to
        capabilities:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
paths:
...
  /gateway/replication:
    get:
      operationId: listReplicationRules
      summary: List replication rules
      tags: [Gateway]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '200': { description: Rules }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      operationId: createReplicationRule
      summary: Create replication rule
      tags: [Gateway]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '201': { description: Created }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /gateway/replication/{id}:
    delete:
      operationId: deleteReplicationRule
      summary: Delete replication rule
      tags: [Gateway]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /site-replication/rules/{id}:
    delete:
      operationId: deleteSiteReplicationRule
      summary: Delete site replication rule
      tags: [Buckets]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '200': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /site-replication/status:
    get:
      operationId: siteReplicationStatus
      summary: Site replication lag and queue
      tags: [Buckets]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '200': { description: Status }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /federation/clusters:
    get:
      operationId: listFederationClusters
      summary: List federation clusters
      tags: [Federation]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
        - name: cluster_id
          in: query
          required: false
          schema: { type: string }
          description: Filter peers by trusted cluster id (local or remote)
      responses:
        '200':
          description: Clusters
          content:
            application/json:
              schema:
                type: object
                properties:
                  clusters:
                    type: array
                    items:
                      $ref: '#/components/schemas/FederationCluster'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      operationId: createFederationCluster
      summary: Register federation cluster
      tags: [Federation]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, endpoint]
              properties:
                name: { type: string }
                endpoint: { type: string }
                region: { type: string }
                cluster_id:
                  type: string
                  description: Trusted cluster id this peer belongs to (defaults to local cluster id)
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  cluster:
                    $ref: '#/components/schemas/FederationCluster'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /federation/clusters/{id}:
    delete:
      operationId: deleteFederationCluster
      summary: Remove federation cluster
      tags: [Federation]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
          - name: id
            in: path
            required: true
            schema: { type: string }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /clusters:
    get:
      operationId: listTrustedClusters
      summary: List local and trusted remote clusters
      tags: [Clusters]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      responses:
        '200': { description: Clusters }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /clusters/{id}:
    get:
      operationId: getTrustedCluster
      summary: Trusted cluster detail and cert timeline
      tags: [Clusters]
      security:
        - BearerJWT: []
        - BearerAPIToken: []
      x-required-role: administrator
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Detail }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...